requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.20.0
attrs==25.4.0
## !! Could not determine repository location
-e /home/hossam/BotCultureGV3
discord.py==2.6.4
frozenlist==1.8.0
greenlet==3.3.0
idna==3.11
iniconfig==2.3.0
multidict==6.7.0
packaging==25.0
pluggy==1.6.0
propcache==0.4.1
pytest==8.3.4
pytest-asyncio==0.24.0
python-dotenv==1.2.1
ruff==0.8.4
SQLAlchemy==2.0.45
typing_extensions==4.15.0
yarl==1.22.0
